+++ packages/stores/src/account/__tests_e2e__/lock.spec.ts
@@ -1,12 +1,12 @@
 /* eslint-disable */
-import { TestOsmosisChainId } from "../../__tests_e2e__/mock-data";
+import { TestOsmosisChainId } from "../../tests/mock-data";
 import {
   deepContained,
   getEventFromTx,
   initAccount,
   RootStore,
   waitAccountLoaded,
-} from "../../__tests_e2e__/test-env";
+} from "../../tests/test-env";
 
 describe("Lock Token Tx", () => {
   const { accountStore } = new RootStore({
