+++ packages/stores/src/tests/mock-keplr-with-fee.ts
