+++ packages/stores/src/tests/setup-tests.ts
@@ -2,8 +2,13 @@
 import "@testing-library/jest-dom";
 import "whatwg-fetch";
 
-import { server } from "./src/__tests_e2e__/msw-server";
+import { server } from "./msw-server";
 
+/**
+ * We use MSW to mock API requests at a network level in non-E2E Jest tests.
+ * This is a global setup for MSW.
+ * @see https://v1.mswjs.io/docs/
+ */
 beforeAll(() => server.listen({ onUnhandledRequest: "warn" }));
 afterEach(() => server.resetHandlers());
 afterAll(() => server.close());
