+++ packages/stores/src/tests/start-osmosisd.sh
