+++ packages/stores/src/tests/stop-osmosisd.sh
