+++ packages/stores/src/tests/test-wallet.ts
