+++ packages/utils/jest.config.js
@@ -4,7 +4,7 @@ module.exports = {
   testMatch: ["**/__tests__/?(*.)+(spec|test).[jt]s?(x)"],
   testEnvironment: "jest-environment-jsdom",
   testTimeout: 100000,
-  setupFilesAfterEnv: ["<rootDir>/setup-tests.ts"],
+  setupFilesAfterEnv: ["<rootDir>/src/tests/setup-tests.ts"],
   watchPlugins: [
     "jest-watch-typeahead/filename",
     "jest-watch-typeahead/testname",
